requirements.txt
Flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
pandas==2.2.1
Flask-WTF==1.2.1
WTForms==3.1.2 